Add WORKFLOW_BACKEND_HOST to code-executor deployment
If retool workflows are enabled, `code-executor` deployments need to know which workflow backend to talk to. Prior to this update, users of the chart would have to manually inject this env var.

This commit adds the (conditional) environment variable `WORKFLOW_BACKEND_HOST` to the `code-executor` deployment in the same way that it is being added in the other deployment manifests.