Fix so CODE_EXECUTOR_INGRESS_DOMAIN is only added if workflows are enabled
When workflows.enabled is set to false, the code executor will not be
deployed. The CODE_EXECUTOR_INGRESS_DOMAIN environment variable would
still be added to backend deployment and prevent it from starting as the
backend would try to connect to the executor on start up.

With this change, the CODE_EXECUTOR_INGRESS_DOMAIN environment variable
will only be added to the backend deployment if workflows are enabled.